agentes-iesjandula/requirements.txt
# --- Núcleo de la API y Servidor ---
fastapi
uvicorn
python-dotenv
gradio
pydantic

# --- Framework de IA y Orquestación ---
langchain
langchain-core
langchain-community
langgraph
langchain-ollama
langchain-google-genai

# --- Herramientas de Búsqueda y Navegación ---
playwright
langchain-tavily
beautifulsoup4
requests
rapidfuzz

# --- Base de Datos Vectorial (RAG) ---
langchain-chroma
langchain-huggingface
sentence-transformers
pypdf
langchain-text-splitters

# --- Utilidades ---
typing-extensions